test: Separate unit from integration tests (#162)
