+++ .circleci/config.yml
@@ -1,8 +1,19 @@
+defaults: &defaults
+  docker:
+    - image: circleci/node:8
+
+git-login: &git-login
+  name: Setting up git user
+  command: git config --global user.email circleci@circleci && git config --global user.name CircleCI
+
+npm-login: &npm-login
+  name: Logging in to npm
+  command: echo "$NPM_TOKEN" > ~/.npmrc
+
 version: 2
 jobs:
-  build:
-    docker:
-      - image: circleci/node:8
+  install-deps:
+    <<: *defaults
     steps:
       - checkout
       - restore_cache:
@@ -12,13 +23,7 @@ jobs:
           name: Installing npm dependencies
           command: npm install
       - run:
-          name: Setting up deployment environment
-          command: |
-            echo "Setting up git user"
-            git config --global user.email circleci@circleci && git config --global user.name CircleCI
-
-            echo "Logging in to npm"
-            echo "$NPM_TOKEN" > ~/.npmrc
+          <<: *git-login
       - run:
           name: Handling package-lock changes
           command: |
@@ -40,50 +45,45 @@ jobs:
           key: v1-npm-deps-{{ checksum "package-lock.json" }}
           paths:
             - node_modules
+      - persist_to_workspace:
+          root: .
+          paths:
+            - .
 
-      # Code Quality checks
-      # TODO: Lint commit message
-      # with git log --format=%B -n 1 $CIRCLE_SHA1 | npx commitlint
-
-      # Build
+  build:
+    <<: *defaults
+    steps:
+      - attach_workspace:
+          at: .
       - run:
           name: Transpiling source code
           command: npm run compile
+      - persist_to_workspace:
+          root: .
+          paths:
+            - out
+
+  build-docs:
+    <<: *defaults
+    steps:
+      - attach_workspace:
+          at: .
       - run:
           name: Creating API documentation
           command: npm run docs
+      - persist_to_workspace:
+          root: .
+          paths:
+            - docs/api
 
-      # Test
+  lint-source:
+    <<: *defaults
+    steps:
+      - attach_workspace:
+          at: .
       - run:
           name: Linting source files
           command: npm run lint -- --format junit --output-file ~/reports/eslint.xml
-          when: always
-      - run:
-          name: Running tests
-          command: node_modules/.bin/nyc npm test -- --reporter mocha-circleci-reporter
-          environment:
-            NODE_ENV: test
-            MOCHA_FILE: ../reports/mocha.xml
-          when: always
-      - run:
-          name: Checking documentation for broken links
-          command: npm run test:docs
-          when: always
-
-      # Coverage checks
-      - run:
-          name: Checking test coverage
-          command: node_modules/.bin/nyc check-coverage --statements 90 --branches 90 --functions 90 --lines 90
-          when: always
-      - run:
-          name: Reporting test coverage
-          command: node_modules/.bin/nyc report --reporter=lcov > coverage.lcov | node_modules/.bin/codecov
-      - run:
-          name: Check documentation coverage
-          command: npm run docs:coverage
-          when: always
-
-      # Store Results
       - store_test_results:
           path: ~/reports
           when: always
@@ -91,24 +91,153 @@ jobs:
           path: ~/reports
           when: always
 
-      # Deployment
+  unit-tests:
+    <<: *defaults
+    steps:
+      - attach_workspace:
+          at: .
+      - run:
+          name: Running tests
+          command: npx nyc npm run test:unit -- --reporter mocha-circleci-reporter
+          environment:
+            NODE_ENV: test
+            MOCHA_FILE: ../reports/unit-tests.xml
+      - run:
+          name: Checking test coverage
+          command: npx nyc check-coverage --statements 90 --branches 90 --functions 90 --lines 90
+          when: always
+      - run:
+          name: Reporting test coverage
+          command: npx nyc report --reporter=lcov > coverage.lcov | node_modules/.bin/codecov
+      - store_test_results:
+          path: ~/reports
+          when: always
+      - store_artifacts:
+          path: ~/reports
+          when: always
+
+  integration-tests:
+    <<: *defaults
+    steps:
+      - attach_workspace:
+          at: .
+      - run:
+          name: Running tests
+          command: npm run test:integration -- --reporter mocha-circleci-reporter
+          environment:
+            NODE_ENV: test
+            MOCHA_FILE: ../reports/integration-tests.xml
+      - store_test_results:
+          path: ~/reports
+          when: always
+      - store_artifacts:
+          path: ~/reports
+          when: always
+
+  test-docs:
+    <<: *defaults
+    steps:
+      - attach_workspace:
+          at: .
+      - run:
+          name: Checking documentation for broken links
+          command: npm run test:docs
+      - run:
+          name: Check documentation coverage
+          command: npm run docs:coverage
+          when: always
+
+  update-lockfile:
+    <<: *defaults
+    steps:
+      - checkout
       - add_ssh_keys
-      - deploy:
+      - attach_workspace:
+          at: .
+      - run:
+          <<: *git-login
+      - run:
           name: Updating lockfile
           command: git push origin $CIRCLE_BRANCH
-      - deploy:
-          name: Deploy changes
-          command: |
+
+  deploy:
+    <<: *defaults
+    steps:
+      - checkout
+      - add_ssh_keys
+      - attach_workspace:
+          at: .
+      - run:
+          <<: *git-login
+      - run:
+          <<: *npm-login
+      - run:
+          name: Deploying changes
+          command:
             npx bump-version release --release-files out docs/api --gh-token $RELEASE_GITHUB_TOKEN
 
-            if [ $CIRCLE_BRANCH = 'master' ]; then
-              echo "Update GitHub pages"
-              version=$(node -e "console.log(require('./package.json').version)")
-              git clone git@github.com:$CIRCLE_PROJECT_USERNAME/$CIRCLE_PROJECT_REPONAME.git -b gh-pages ../pages
-              cp -r docs/api/* ../pages
-              cd ../pages
-              git add .
-              git commit -m "docs(gh-pages): Update for release v$version [ci skip]"
-              git push origin gh-pages
-              cd ../project
-            fi
+  update-hosted-docs:
+    <<: *defaults
+    steps:
+      - checkout
+      - attach_workspace:
+          at: .
+      - run:
+          <<: *git-login
+      - run:
+          name: Updating hosted API documentation
+          command: |
+            version=$(node -e "console.log(require('./package.json').version)")
+            git clone git@github.com:$CIRCLE_PROJECT_USERNAME/$CIRCLE_PROJECT_REPONAME.git -b gh-pages ../pages
+            cp -r docs/api/* ../pages
+            cd ../pages
+            git add .
+            git commit -m "docs(gh-pages): Update for release v$version [ci skip]"
+            git push origin gh-pages
+            cd ../project
+
+workflows:
+  version: 2
+
+  test-and-deploy:
+    jobs:
+      - install-deps
+      - build:
+          requires:
+            - install-deps
+      - build-docs:
+          requires:
+            - install-deps
+      - lint-source:
+          requires:
+            - install-deps
+      - unit-tests:
+          requires:
+            - install-deps
+      - integration-tests:
+          requires:
+            - build
+      - test-docs:
+          requires:
+            - build-docs
+      - update-lockfile:
+          requires:
+            - lint-source
+            - unit-tests
+            - integration-tests
+            - test-docs
+          filters:
+            branches:
+              only: /greenkeeper\//
+      - deploy:
+          requires:
+            - lint-source
+            - unit-tests
+            - integration-tests
+            - test-docs
+      - update-hosted-docs:
+          requires:
+            - deploy
+          filters:
+            branches:
+              only: master
